requirements.txt
blinker==1.4
Bootstrap-Flask==2.0.2
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-DebugToolbar==0.13.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
gunicorn==20.1.0